agronomia-master/requirements.txt
Pillow==5.3.0
django==2.1.3
django-ckeditor==5.6.1
django-filter==2.1.0
wagtail==2.4
django-social-share==1.3.2


